Hello People, We just finish loading and using mutiple animation, with a single character, Now we can load a load of animation and and apply it to a skinned mesh, and it will work like a acharm, and now the asset manager manage the animation we can query animation from asset amanager, and we fixed the weird glich with strach the trainagle, eee what else...... and next it i think iterpolation b/n two animation we will see how it goes... until then peace ooooott

+++ AssetLoader/src/Geometry/CoreGeometryLoader.cpp
@@ -252,12 +252,12 @@ namespace OE1Core
 						float weight = bone_weights[k].mWeight;
 
 						// set vertex with the weights
-						for (size_t j = 0; j < MAX_BONE_PER_VERTEX; j++)
+						for (int j = 0; j < MAX_BONE_PER_VERTEX; j++)
 						{
-							if (unprocessed_geom.VertexData[vertex_id].BoneIndex[(int)j] < 0)
+							if (unprocessed_geom.VertexData[vertex_id].BoneIndex[j] < 0)
 							{
-								unprocessed_geom.VertexData[vertex_id].BoneIndex[(int)j] = (int)BoneID;
-								unprocessed_geom.VertexData[vertex_id].Weight[(int)j] = weight;
+								unprocessed_geom.VertexData[vertex_id].BoneIndex[j] = BoneID;
+								unprocessed_geom.VertexData[vertex_id].Weight[j] = weight;
 								break;
 							}
 						}
